requirements.txt
# FastAPI 및 웹 프레임워크
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3
pydantic==2.10.5
pydantic_core==2.27.2

# 데이터베이스
sqlalchemy
pymysql
pymongo[srv]
python-dotenv

# 데이터 처리 및 분석
pandas
numpy
scikit-learn

# HTTP 클라이언트 (테스트용)
httpx

# 유틸리티
requests
email-validator==2.2.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0

# 타입 힌트 및 검증 (호환성 해결)
annotated-types==0.7.0
typing_extensions==4.14.0

# 비동기 처리
anyio==4.8.0
sniffio==1.3.1

# HTTP 관련
h11==0.14.0
idna==3.10

# 개발 도구
mypy==1.14.1
mypy-extensions==1.0.0
tomli==2.2.1

# 예외 처리
exceptiongroup==1.2.2

# 자연어 처리 (호환성 해결)
nltk>=3.9
textblob>=0.19.0
tokenizers>=0.21,<0.22

# RAG 시스템 필수 패키지
faiss-cpu
langgraph
streamlit
langchain
langchain-core